backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
langchain==0.1.12
langchain-community==0.0.28
langchain-groq==0.1.1
faiss-cpu==1.7.4
numpy==1.26.4
pandas==2.2.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
mysqlclient==2.2.4
python-jose==3.3.0
pydantic==2.6.3 